requirements.txt
streamlit
pandas
openpyxl
plotly
pytz
gspread
google-auth
google-api-python-client
